Fix bug with definition of parameter maximum cycle driver: change double -> int Change: now computation of dt in "flow and reactive transport" is based on Flow PK only.

+++ src/common/interface_platform/InputParserIS_MPC.cc
@@ -558,7 +558,7 @@ Teuchos::ParameterList InputParserIS::CreateCycleDriver_List_(Teuchos::Parameter
 
   cycle_driver_list.sublist("time periods").sublist(tp_list_name.data()).set<double>("start period time", switch_time);
   cycle_driver_list.sublist("time periods").sublist(tp_list_name.data()).set<double>("end period time", end_time);
-  cycle_driver_list.sublist("time periods").sublist(tp_list_name.data()).set<double>("maximum cycle number", max_cycle_number);
+  cycle_driver_list.sublist("time periods").sublist(tp_list_name.data()).set<int>("maximum cycle number", max_cycle_number);
   cycle_driver_list.sublist("time periods").sublist(tp_list_name.data()).set<double>("initial time step", dt_tran);
   cycle_driver_list.sublist("Time Period Control") = tpc_list;
